requirements.txt
fswrap==0.1.1
commando==0.3.2a
PyYAML==3.10
Markdown==2.3.1
MarkupSafe==0.15
Pygments==1.6
typogrify==2.0.0
Jinja2==2.6
